remove explicit replication check via md5sum
First, it was prone to a race condition where if something is changed
between the time the database is imported into the slave and the time
the next slave backup is done, then md5sum doesn't match, but the
replication was actually fine.

Also, we are now importing the data into the slave inside a single
transaction, so if it fails, then the slave will already be in its
previous state.

Also, it was currenly not working since the slave backup can't be
imported directly; the truncate statement must be emitted first.