force cc_agent_policy to never on forward
What happens if it is set to generic is that the caller could then
request call completion, but it might not work as he expected.

An example is a user A with a schedule that is calling user B; if
someone call user A, user B will ring. If the caller then request call
completion, it will monitor B's phone, but when B becomes available, it
will dial user A extension.