force caller ID on callee channel in outgoing calls
Ideally, we would like to set the caller ID number on the callee channel
to the caller XIVO_BASE_EXTEN *only* if the caller ID on the callee
channel is not already set, but the Dial application doesn't have such
an option.

We could also use pre-dial handler to do such a thing, but since
pre-dial handler are limited to 1, this would break the dialplan of
people who are already using pre-dial handler in their outgoing call
subroutines.

Adding this logic directly in XiVO would take quite some time to do it
right, but I'm not sure this is worth it. Especially that it doesn't
feel to be the right place to do such a thing.

Ticket #5036.