Update extensions_lib_subr.conf
This a feature request from many customers. The need is this integration of the destination number in the fax file naming.

In case of Fax to FTP, if the FTP repository is unavailable all the pending fax are mixed into the /var/spool/asterisk/fax. With this new naming you are able to retrieve the destination of each fax.