XIVO_QUEUEOPTIONS: use single instead of multiple inheritance
Since it is only needed on the Local channel that is created by
app_queue when calling an agent member. That said, using multiple
inheritance was not causing any problems right now, but if one day the
AGI incmonig_queue_set_features doesn't set inconditionnally the
QUEUE_OPTIONS variable, it could cause one.