vmbox: fallback on transferer language
Before this changes, when doing an attended transfer to a voicemail
using the feature code (like the XiVO client does), then the voicemail
prompt were always in english since the Local channel's language is not
inherited from the transferer channel and was not explicitly set, i.e.
it was not using neither the transferer or transferee language.

With this changes, if the voicemail has no language set and you are in
the "attended transfer from feature code" case, then use the language of
the transferer channel language as the channel's language.