add drop-in file for monit.service
In HA mode, we disable monit with START=no in /etc/default/monit. But
this method is sysV style, not systemd. So when systemd try to
start the minimal xivo-service for the HA, it execute the init.d file
and put this status on Active (exited).

With the option RemainAfterExit=no, systemd will not put this status on
Active, because there are no process associate to monit.