odbc: add pooling to xivo DSN
Asterisk 13.8 has removed ODBC connection pooling in res_odbc, so we
need to enable it in unixodbc for performance reason.

Pooling is enabled per *driver* via the CPTimeout (Connection Pooling
Timeout) option, the value indicating "the number of seconds a pooled
connection will remain open if it is not being used". So to force the
xivo DSN to use connection pooling, add a "xivo" driver and tell the DSN
to use that driver.