readd the xivo-callme context
The context was removed in commit 9183d3a274, but some administrators are
relying on it, so add it back.

The xivo-callme context is used for troubleshooting purpose, e.g. to ring a phone
remotely from the asterisk CLI with:

   channel originate Local/1234@default extension 42@xivo-callme

The file is now part of the "official xivo dialplan" instead of being in
the /etc/asterisk/extensions_extra.d directory.