remove xivo odbc driver
It seems like unixodbc has some trouble interpreting correctly it's
odbcinst.ini file when two drivers with different names are referring
to the same "Driver" (i.e. same .so file). In this case, even if the "xivo"
driver was setting "CommLog = 0", unixodbc was using the "CommLog = 1"
setting of the "PostgreSQL ANSI" driver. This was resulting in asterisk
creating noisy log files in /var/spool/asterisk directory on each
restart.

So remove the xivo driver and just use the PostgreSQL ANSI driver, and
modify the PostgreSQL ANSI driver settings.