xivo_fax.conf: move examples into separate example file
By removing the examples from /etc/xivo/asterisk/xivo_fax.conf, we don't
have to annoy every administrator which have modified theirs xivo_fax.conf
when we just want to update the examples, and when the file do get changed,
it's easier for them to see what has changed.

The example file is actually in the xivo-agid package.