debian: don't depends on xivo-nginx
This is causing an issue during the 14.01 to 17.02 upgrade, because of
xivo-config, xivo-nginx and xivo-web-interface relationships
(pre-depends, depends, breaks).

The dependency is not strictly necessary, i.e. even if xivo-nginx is not
installed, the xivo-config package still works fine (but you don't have
the possibility to connect to asterisk websocket through nginx). Plus,
since xivo-nginx reloads nginx either when triggered or in its postinst,
the nginx configuration should still be reload and fine at the end of
the upgrade.