xivo-create-config: force dhcp_integration
reason: it cant be configured from confd and it does not cause any
trouble when always enabled.

If it ever causes trouble and we want to implement the option in confd,
the previous value is still stored in the database.  If it proves to be
good enough to leave it always enabled, we may then remove the value
from the database.