user: remove the wake mobile in the pre-dial hook
this implementation will block the call for up to "timeout" seconds before
entering the dial.

this delay happens when the a mobile session exists for a given user but there
are no mobiles to wake.

When that happens the call will not start for a LONG time (usually 30 seconds)
event if the user has multiple devices which are already registered