Fix issue with some calls are very long
Sometimes on webrtc we have some calls who are very long. After some research we find there is an issue with the buffer size on ssl termination  with nginx. This issue is not present with apache, but in fact we don't need to have an ssl termination between asterisk and nginx by default. So just change the configuration to use the asterisk http endpoint instead https fix this issue.