voicemail: on destination voicemail do not ignore the options
In this context ARG2 contains options for the Voicemail application. The
possible values are "b" and "u" which are mutually exclusive.

When "b" is used the message indicates that the user is busy.
When "u" is used the message indicates that the user is unavailable.
When none of the above is set the we are invited to leave a message.

So, whats happening here.

FILTER(u,$(ARGS)) will return "u" if ARG2 contains "u" ex: "busy" ==> "u"
The IF will match the "u" and set the XIVO_MAILBOX_OPTIONS to "u"

Since "b" is mutually exclusive with "u" we can safely overrite the previous
value if "b" is set.