remove the XIVO_CALLRECORDFILE variable
This variable has been replaced by the

WAZO_CALL_RECORD_FILE_CALLEE
WAZO_CALL_RECORD_FILE_CALLER
WAZO_CALL_RECORD_SIDE

variables.

Having only one variable does not work since we are trying to record both sides
of a call.

The WAZO_CALL_RECORD_SIDE variable contains "caller" or nothing. This variable
is not inherited allowing any further code to known if it's the caller or the
callee if the variable does not exist. From that information the appropriate
variable can be looked up to find which filename to use for the recording