user: make ARG2 of user more dynamic
ARG2 can now contain the ringtime as well as a MOH UUID, positions are important
here. If the ringtime is undefined ARG2 SHOULD be ";moh_uuid"

the AGI incoming_user_set_features is then responsible for finding the MOH based
on the UUID and add the "m" option for the "Dial" application.