pickup: add a wazo-pickup subroutine
The Pickup application does not work well with Stasis applications. Adding some
dialplan around this application allows us to do specific routing before doing
the Pickup.

In this case we are going to try to pickup the any dial_mobile that have been
started on a AOR matching the PICKUPMARK we are picking up from.